+++ src/LedMatrixPatterns.h
@@ -17,7 +17,7 @@ extern byte num5[8];
 extern byte smile[8];
 extern byte timedout[8];
 
-void initDisplay(LedController &lc, uint8 intensity);
+void initDisplay(LedController &lc, unsigned int intensity);
 void loadingAnimation(LedController &lc);
 void writeMatrix(LedController &lc, byte bname[8]);
 void writeMatrixInv(LedController &lc, byte bname[8]);
